# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Room 911",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://servicios.asstiseguridadsocial.com"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Crear usuario administrador",
                "operationId": "3e0d6c1421f70b3e11ce33c629ecb9d3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Successfully created user!"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/authentication": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ingreso de usuarios",
                "operationId": "c757f9366ac8fe96b9b957998b54cb98",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "schema": {
                            "type": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "access_token": "Acceso correcto",
                                        "token_type": "",
                                        "expires_at": "",
                                        "user": []
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Listar departamentos",
                "operationId": "74d96038b7f2328679066615b851ca5f",
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "department": []
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Crear departamento",
                "operationId": "eea1c299672895cbb560b0f89b6375f4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Departamento creado con éxito"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/departments/{id}": {
            "put": {
                "tags": [
                    "Departments"
                ],
                "summary": "Actualizar departamento",
                "operationId": "c05be8f7715d3015bc5bfa1c6ea6fd49",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Departamento creado con éxito"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}